# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.Response:
    properties:
      data: {}
      error: {}
      extra: {}
      message:
        type: string
      name:
        type: string
      pagination: {}
      status:
        type: string
      status_code:
        type: integer
    type: object
...
  firs_models.FirsTransactionVatPayload:
    properties:
      agentTin:
        type: string
      baseAmount:
        type: string
      beneficiaryTin:
        type: string
      currency:
        type: integer
      itemDescription:
        type: string
      otherTaxes:
        type: string
      totalAmount:
        type: string
      transDate:
        type: string
      vatCalculated:
        type: string
      vatRate:
        type: string
      vatStatus:
        type: integer
      vendorTransactionId:
        type: string
    type: object
...
  firs_models.VerifyTinData:
    properties:
      tin:
        type: string
    required:
    - tin
    type: object
...
info:
  contact: {}
  description: This is the e-invoicing service API documentation.
  termsOfService: http://swagger.io/terms/
  title: Gention E-invoicing Service API
  version: "1.0"
paths:
...
  /entity:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of entities
      parameters:
      - in: query
        name: page
        type: integer
      - in: query
        name: reference
        type: string
      - in: query
        name: size
        type: integer
      - in: query
        name: sortBy
        type: string
      - in: query
        name: sortDirectionDesc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Entities retrieved successfully
          schema:
            $ref: '#/definitions/entities.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/entities.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.Response'
      security:
      - BearerAuth: []
      summary: Get Entities
      tags:
      - Entity
  /entity/{entity_id}:
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific entity using its ID
      parameters:
      - description: Entity ID
        in: path
        name: entity_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Entity retrieved successfully
          schema:
            $ref: '#/definitions/entities.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/entities.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.Response'
        "404":
          description: Entity not found
          schema:
            $ref: '#/definitions/entities.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.Response'
      security:
      - BearerAuth: []
      summary: Get Entity by ID
      tags:
      - Entity
  /entity/vat-payment:
    post:
      consumes:
      - application/json
      description: Submit VAT payment details to FIRS
      parameters:
      - description: VAT payment request payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/firs_models.FirsTransactionVatPayload'
      produces:
      - application/json
      responses:
        "200":
          description: VAT payment processed successfully
          schema:
            $ref: '#/definitions/entities.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/entities.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.Response'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/entities.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.Response'
      security:
      - BearerAuth: []
      summary: Post VAT Payment
      tags:
      - Entity
  /entity/verify-tin:
    post:
      consumes:
      - application/json
      description: Verify a taxpayer identification number (TIN)
      parameters:
      - description: TIN verification request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/firs_models.VerifyTinData'
      produces:
      - application/json
      responses:
        "200":
          description: TIN verified successfully
          schema:
            $ref: '#/definitions/entities.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/entities.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.Response'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/entities.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.Response'
      security:
      - BearerAuth: []
      summary: Verify TIN
      tags:
      - Entity
...
  /invoice/transmit/{irn}:
    post:
      consumes:
      - application/json
      description: Transmits an invoice to FIRS using the IRN.
      parameters:
      - description: Invoice Reference Number (IRN)
        in: path
        name: irn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice transmitted successfully
          schema:
            $ref: '#/definitions/entities.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/entities.Response'
      security:
      - BearerAuth: []
      summary: Transmit Invoice
      tags:
      - Invoice
  /invoice/transmit/confirm/{irn}:
    get:
      consumes:
      - application/json
      description: Confirms a transmitted invoice using the IRN.
      parameters:
      - description: Invoice Reference Number (IRN)
        in: path
        name: irn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice confirmed successfully
          schema:
            $ref: '#/definitions/entities.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/entities.Response'
      security:
      - BearerAuth: []
      summary: Confirm Transmitted Invoice
      tags:
      - Invoice
...
  /invoice/transmit/lookup-irn/{irn}:
    get:
      consumes:
      - application/json
      description: Retrieves invoice details using the IRN.
      parameters:
      - description: Invoice Reference Number (IRN)
        in: path
        name: irn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice details retrieved
          schema:
            $ref: '#/definitions/entities.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/entities.Response'
      security:
      - BearerAuth: []
      summary: Look Up IRN
      tags:
      - Invoice
  /invoice/transmit/lookup-party/{party_id}:
    get:
      consumes:
      - application/json
      description: Retrieves details using Party ID.
      parameters:
      - description: Party ID
        in: path
        name: party_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Party ID details retrieved
          schema:
            $ref: '#/definitions/entities.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/entities.Response'
      security:
      - BearerAuth: []
      summary: Look Up Party ID
      tags:
      - Invoice
  /invoice/transmit/lookup-tin/{tin}:
    get:
      consumes:
      - application/json
      description: Retrieves taxpayer details using TIN.
      parameters:
      - description: Tax Identification Number (TIN)
        in: path
        name: tin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: TIN details retrieved
          schema:
            $ref: '#/definitions/entities.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/entities.Response'
      security:
      - BearerAuth: []
      summary: Look Up TIN
      tags:
      - Invoice
  /invoice/transmit/pull:
    get:
      consumes:
      - application/json
      description: Pulls invoices from FIRS using query params.
      parameters:
      - in: query
        name: confirmed
        type: string
      - in: query
        name: from
        type: string
      - in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoices pulled successfully
          schema:
            $ref: '#/definitions/entities.Response'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/entities.Response'
      security:
      - BearerAuth: []
      summary: Pull Transmitted Invoices
      tags:
      - Invoice
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"